+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bugfixes
 - [#2282](https://github.com/influxdb/influxdb/pull/2282): Use "value" as field name for OpenTSDB input.
 - [#2283](https://github.com/influxdb/influxdb/pull/2283): Fix bug when restarting an entire existing cluster.
+- [#2293](https://github.com/influxdb/influxdb/pull/2293): Open cluster listener before starting broker.
 
 ## v0.9.0-rc24 [2015-04-13]