+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - [#2276](https://github.com/influxdb/influxdb/pull/2276): Broker topic truncation.
 - [#2292](https://github.com/influxdb/influxdb/pull/2292): Wire up drop CQ statement - thanks @neonstalwart!
 - [#2290](https://github.com/influxdb/influxdb/pull/2290): Allow hostname argument to override default config - thanks @neonstalwart!
+- [#2295](https://github.com/influxdb/influxdb/pull/2295): Use nil as default return value for MapCount - thanks @neonstalwart!
 
 ## v0.9.0-rc24 [2015-04-13]