+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## v0.9.0-rc25 [unreleased]
+## v0.9.0-rc25 [2015-04-15]
 
 ### Bugfixes
 - [#2282](https://github.com/influxdb/influxdb/pull/2282): Use "value" as field name for OpenTSDB input.